+++ CHANGELOG.md
@@ -7,6 +7,20 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 0.7.5 - Unreleased
 
 
+### Added
+
+* New CI for building wheels with cibuildwheel
+* Checkerboard demo image
+* `warp_affine` with antialiasing for downsampling by default
+* `Affine.decompose` which extracts the scale, translation, rotation, and shear
+  from an affine matrix.
+
+
+### Changed
+* `imscale` is deprecated and now results in an error
+* `warp_image`, which did not do what you think it does, is deprecated and now results in an error
+
+
 ## Version 0.7.4 - Released 2021-05-13